[Refactor] ProjectDetail로 이름 변경 & 컴포넌트 분리
