+++ src/components/ProjectDetail/Comment/index.tsx
@@ -1,0 +1,5 @@
+import * as S from '@/components/ProjectDetail/Comment/styles';
+
+export const Comment = () => {
+  return <S.CommentContainer></S.CommentContainer>;
+};
