+++ src/components/ProjectDetail/Comment/styles.ts
@@ -1,0 +1,7 @@
+import styled from 'styled-components';
+
+export const CommentContainer = styled.div`
+  width: 100%;
+  background-color: ${(props) => props.theme.colors.gray5};
+  border-top: 0.1rem solid ${(props) => props.theme.colors.gray20};
+`;
