+++ src/pages/Project/ProjectDetailPage/styles.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+export const Layout = styled.div`
+  width: 100%;
+  margin-top: 6.4rem;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+`;
